abstracted map generation to separate class, further improvements incoming.
